WIP reactions sub-apps migration to 1.0
