+++ reactions/workchains/__init__.py
@@ -1,0 +1,7 @@
+"""on-surface reaction work chains."""
+# pylint: disable=unused-import
+
+from __future__ import absolute_import
+
+from .replicawork import ReplicaWorkChain
+from .nebwork import NebWorkChain
